Move public V3 protocols into separate file
